producerDataCDMX/requirements.txt
APScheduler==3.6.3
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
confluent-kafka==1.5.0
Fiona==1.8.13.post1
geopandas==0.8.1
idna==2.10
kafka-python==2.0.1
munch==2.5.0
numpy==1.19.1
pandas==1.1.1
pygeos==0.7.1
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
Shapely==1.7.1
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
